AzureDevOpsPytest_InfraTestingFramework/Governance-Policies/requirements.txt
pytest
pytest-cov
pytest-azurepipelines
unittest-xml-reporting
azure-mgmt-resource
azure-mgmt-storage
azure-identity